Fix errors with download
Prevents the errors
`mv: cannot move 'pwtmp/.' to './.': Device or resource busy` and 
`mv: cannot move 'pwtmp/..' to './..': Device or resource busy`.

The error occurs because we try to move the special directories "." and ".." which represent the current directory and the parent directory respectively. This operation is not allowed, hence the error.